[FIX] web_editor: fix color after link
When clicking the link tools button, the focus is set inside the URL
input, changing the selection temporarily to that input. This means that
if the user clicks any other button from the toolbar right after using
that input, a command would try (and refuse) to execute outside the
editable area.
This restores the last selection in editable if it exists, whenever we
call a command on a non-editable area.

task-2921844