[IMP] website_hr_recruitment: improve filters
Before this commit, filters were displayed as a set of `nav-pills` items
in a column on the left of the main template.

After this commit, filters have been moved to the top as dropdown 
elements, displaying counts inside badges, which are obtained from the 
list of filtered jobs.

Additionally, `Remote` filterable location has been included, and can
be set either by using the backend or directly with the website builder.

task-2406626

Co-authored with Benoit Socias <bso@odoo.com>