[IMP] mail: removes mail.message inheritance from mail.mail
(EXPLORATION PR)

The goal is to remove the inheritance of mail.message from mail.mail because
it was causing right access problem that was solved by a small hack on the ORM
by adding sudo related field to the parent: email_from, reply_to, subject.

To solve the problem, we would instead make mail.message linked to mail.mail
by a many2one relationship and add fields used as related fields.

Technical note: We have 2 options to implements it:
- either we modify all calls to create and write, splitting the call in 2: one
for mail.message and one for mail.mail. But we have to identify everywhere it
is used
- or we simulate the inheritance for the create and write methods by
transmitting the values "to the mail.message"

Task-2859898