[IMP] web: only enable spellcheck on focused inputs
This commit adds the ability to enable spellcheck on textarea
only if they are in focus. This avoids the UI to be bloated
with unwanted underlined text on unfocused elements when the
upcoming in a context with always edit mode.

A core hook has been introduced to allow other components to
use the same behavior without reimplementing it. It is currently
used by the TextField. It means it is also used in the HtmlField.

task #2861428