[REF] website: review options for the dynamic snippet
Prior to this commit, the Dynamic Snippet options would modify the DOM
at places were it's not supposed to. This commit aims at making the
option follow the flow of an option:

- Modify the target's DOM either on build or on a change of option
- Modify the UI either during willStart or updateUI

To do so, the methods that fetch the template data needed to be split
form the methods that create the buttons.

The benefits are that the option is less likely to cause race condition
and errors while performing its task.

task-2677203