[IMP] mrp(_account): byproduct valuation fix for flexible consumption
PURPOSE:

1. For products that are manufactured with FIFO/AVCO product categories, once a
production is marked as done, users are given the possibility to unlock the MO
and update either
- the byproduct quantity : in the case the component consumption remains
unchanged

SPECIFICATION:

1. Modify By-Product Production

1. Use case 1 : Modify Byproduct

Example :
Produce 5 Litres worth $10.
Get 5 Litres (byproduct)
Modify byproduct to 6 Litres.

Current Behaviour : Post 1 new Layer, for 1 L worth $2

To do Instead :
    Post 0 quantity layer at - $2
    Post 1 quantity layer at + $2

And we make sure that the layers are linked so that the quantities are linked to
the original $10 Layer.

Task - 2574613