[FIX] web: allow quick-creating column with enter in kanban
With the rewrite of the kanban view, the ability to quickcreate a kanban
column by pressing enter has been lost. This is due to the fact that the
hotkey service protects editable areas from hotkeys by default.

This commit fixes that by bypassing the editable protection when
registering the enter hotkey (the escape hotkey bypasses it by default).