[ADD] l10n_ke_edi_copycat, hw_l10n_ke_edi_copycat
This commit adds two modules in order to support Kenyan EDI.
The Kenyan EDI system works in conjunction with a physical device. In
this case the physical device is the Novitus Deon, and the API used to
communicate with the device is provided by CopyCat.

The module is named copycat since we might choose to support other APIs
and other fiscal devices in the future.

l10n_ke_edi_copycat:

This module contains the logic for exporting an invoice in accordance
with the json format expected by the API. The posting of the invoice is
done via a button that triggers a client action. The client action is
defined in javascript, and communicates with a local proxy server, which
in turn communicates with the fiscal device. This is necessary as the
device will be on the client' local network, and it is not guaranteed
that the user is running odoo on premise.

The response from the fiscal device should contain, amongst other
details, a link to the record of the invoice on the Kenyan government
site and a unique code for the invoice. Some of the details from the
response are displayed at the bottom of the invoice.

hw_l10n_ke_edi_copycat:

This module simply contains a controller with one endpoint. The purpose
of this endpoint is to take the content of the request, inside which is
the address of the fiscal device on the local network, and post the
content of request to the address of the fiscal device. The reason for
this is to avoid problems relating to cross origin resource sharing. It
could be run locally, or on an IoT box, with the --load argument, the
address of the server running this module, and the fiscal device that it
forwards to should be stored in fields on the company as defined in
l10n_ke_edi_copycat.

task-id: 2950308

Co-authored-by: jco-odoo <jco@odoo.com>