[IMP] web, base: do not store the model of the relational properties
Purpose
=======
Do not store the model of the many2one / many2many properties on each
child.

Now, to reset the value on the child if the model has been change, we
generate a new name for the property. So it's like if we removed the
property and create a new one.

To be able to restore the old value (e.g. if by mistake we changed the
model, and go back to the old model), we store the initial states.

Task-2852259