[FIX] web: mute Kanban cards with progressbar filter
This commit is a fix to bring back the legacy behavior
in a Kanban view which muted a card when using filter
if the progress has just been changed. It is greyed
out to indicate that it is no longer related to the
filter instead of just staying in the list. A test has
been added as well.