[IMP] base: improve the performance of properties when used in batch
Purpose
=======
Check the existence of the relational properties (many2one / many2many)
in batch and prefetch the values in batch as well to reduce the number
of SQL queries.

Technical
=========
The existence is checked in the read method of the properties field,
because we have the hole recordset. Then, the non-existing ids are
remove from the properties values and the cache is updated.

The dirty flag is set to True if the value has been changed, it means
that now the value can be updated in database, if some records have
been deleted, only by reading the field.

It has the advantage to have a "cleaner" database (the properties value
are cleaned more often), we can trust the cached value (no need to
check for record existence each time we read the cache), but it has the
inconvenient to add a SQL query the next time we read the field,
if one of the records has been deleted (`dirty=True`).

This trade off is based on the fact that we should have way more
read in batch, than records deleted in the database.

Task-2965523