[IMP] base: search on properties fields
Purpose
=======
Allow to search properties fields.

For now, the properties fields can be searched only in the top bar,
and we can not search on relational properties display name
(`[('properties.partner_id', 'ilike', 'bob')]`).

The "in" operator is used to search tags and many2many. Because we can
not know the property type, and because the "in" operator is reserved
to search into JSON array, we can not used domain like
`[('properties.my_char', 'in', ['a', 'b'])]`.

Task-2980121