[IMP] base: search on properties
Purpose
=======
Allow to search properties.

For now, the property can be searched only in the "custom filter",
and we can not search on relational properties display name
(`[('properties.partner_id', 'ilike', 'bob')]`).

The "in" operator is used to search tags and many2many. Because we can
not know the properties type, and because the "in" operator is reserved
to search into JSON array, we can not used domain like
`[('properties.my_char', 'in', ['a', 'b'])]`.

Task-2980121