[IMP] web_editor, website: allow/forbid grid mode for each snippet
- All the snippets that are allowed to have the grid mode now have both
the option "Layout Grid/Cols" and the ability to toggle on drag.

- The snippets who cannot toggle the grid mode have the move handle
back but it doesn't toggle on drag and they don't have the option.

- The drag and drop has been improved to allow normal columns to go
inside the grids.

- Items has been improved since it didn't work well in grid because of
its "hardcoded" height.

- Carousel has been improved by deleting the vertical dropzones
appearing when dragging inner content because they appeared in grid
mode and because now that it has a column option, additional content
can be added by using it.