[FIX] web: open m2o record in target current
Since the readonly mode is removed from the form view, the m2o record
opens in target new (dialog) as we are in edit mode.
With this commit, it opens the record in target current (fullscreen).
This commit also introduce an option to open the record in target new.

task id: 2822553