[IMP] web_editor: solve the jump issue when dragging a grid item
The dropzones appearing in some snippets make their columns go to the
next line, making all other elements shift downwards. When dragging a
grid item from a grid located below these snippets, it results in the
grid jumping downwards, and the grid item is therefore not over its
grid anymore, which is unpleasant.

This commit solves this issue by scrolling in a way that the grid is at
the center of the screen if this situation happens.