[FIX] web: fix crash when drag & dropping in group by date
Drag and dropping a record from one group to another updates some data
inside the groups themselves however they relied on the value of the
field that is being grouped by.
However the ORM does not fill values for date fields in read_groups (it
only contains a formatted value) so the webclient crashed when comparing
dates to undefined values.
This commit fixes that behaviour.