[FIX] web: correctly allow to leave form view
This commit fixes two issues occurring when the user tries to
leave a form view.

Issue 1: open a record that is invalid in database (a required
field is empty), and click on the pager to go to another record.
Before this commit, we prevented the user from leaving the record,
marked the field as invalid and displayed a notification. This is
a weird UX experiment, since the user changed nothing (he did not
even enter edit mode). In this situation, since the record hasn't
been modified, the user should be allowed to leave. This is how
the legacy form view behaved.

Issue 2: open a record, make a change in an input field, and
go back using the breadcrumb **with the hotkey** (alt+b), without
bluring the input. Before this commit, the change was lost.