[FIX] web_editor: fix bug when over and out don't follow each other
When going from a dropzone to an other, and if it is done really fast
or if the dropzones are really close (glued to each other), the order
of the "over" and "out" dropzone events is not the one expected.
Indeed, instead of having over > out > over, we have over > over > out.
This resulted in a traceback when using the grid as the listeners and
properties were therefore not removed correctly when such a case
happened.

This commit adds the the support for this case.