[REF] message_post expect body as Markup
message_post now expect the body of the message to be an instance of Markup.
In case it is not an instance of Markup, escape the content.

This is done to avoid being able to inject html in lots of the usage of message_post

Most of the calls of message_post that needed the body to be in html before are modified to use Markup.