[FIX] hr: incorrect reference used for work address of admin employee
Since commit [1], we use base.main_company as default value for the
address of the admin employee. This is not correct as base.main_company
is referring to res.company model while address_id is expecting a
res.partner id.

This works in most default situations as id = 1 is the same for both
res.company and res.partner, but not anymore in case we have merged
the initial partner with another one.

We should use base.main_partner instead of base.main_company.

[1]:https://github.com/odoo/odoo/commit/c0ca736193df42814389151956eb07c28c77d18e

closes odoo/odoo#100785

Signed-off-by: Kevin Baptiste <kba@odoo.com>