[IMP] web: no min height for oe_title h tags
There was no apparent reason than choice, and we've changed that
choice.
Indeed, you sometimes want so field to be immediately associated
with the title h1 (usually a name).

In this commit, we also introduce the change to led to this choice:
we want a contact's company to be more noticeable and more closely
associated with the name of the contact itself.

Task-2992695