[IMP] website: do not use loading="lazy" on the logo
Just to improve not having useless attributes on the logo image of the
navbar, we remove the loading="lazy" which is automatically added (it
is useless as the logo is very likely to be rendered always anyway and
should be very small).

This solution might be overkill though. The automatic lazy loading is
probably meant to be reviewed at some point anyway.