change name of hsn codes -> hs
throughout the documentation, harmonized system codes are refered to as
hs codes. It would be mildly confusing for those reading the
documentation each time, so the attribute names and references to the
attributes have been adapted as such:
l10n_ke_hsn_code -> l10n_ke_hs_code
l10n_ke_hsn_name -> l10n_ke_hs_description (as it is refered to this way
in the documentation)

_l10n_ke_remove_special_chars has had the length parameter removed
It makes much more sense to me to adapt the length of the string by
slicing it where it is used.
e.g. _l10n_ke_remove_special_chars(whatever_string)[:40] is more
transparent
Also the length parameter failed to take
into account that we might want the end of the string.
e.g. _l10n_ke_remove_special_chars(whatever_string)[-15:]

relevant invoice number has been adapted to reference the correct
attribute of the credit note (it is used in the case of credit notes,
and maybe we will support it for debit notes later on). The attribute
the field should reference is the middleware_number field of the
previous invoice, not the name