[IMP] test_lint: SQL injection reduce false positive Change SQL injection detection of variable type
Previously, a query using a "%" or "+" operator was flagged as injectable if one of the operands was not a constant value.

For example:

cr.execute('SELECT * FROM table WHERE id = %s' % '45') was fine but

var1='45'
cr.execute('SELECT * FROM table WHERE id = %s' % var1)
was flagged as injectable.

However, the value of var1 was already inferred as a constant by
checkers.utils.safe_infer(node)

This result seems robust and could be used to reduce the false positives of the SQL_injection test.