[IMP] web: prevent users from leaving without saving
Before the switch to "always edit" on form views, there was a mechanism
to save "urgently" - i.e. throw everything at the server and hope for
the best, since you cannot wait for the reponse (the window is being
closed and async is out of the picture).

However, this is no longer a behaviour we wish to have - with the always
edit mode, it is more likely that users will want to be aware that there
are unsaved changes before leaving - and it more easily allow users to
handle or notice invalid data

This commit remove all the "urgent save" plumbing and replaces it with a
simple event handler for the beforeUnload event - if the record is dirty
(or new and as of yet unsaved), the browser will ask for confirmation
using its own standard dialog.

Similar changes have been done on editable lists to have a single
coherent behaviour across views with regards to pending changes.