[IMP] analytic: batch calls to server
the analytic distribution json only contains the analytic account id and percentage - and therefore requires a call to the server to fetch additional information such as the name and colour.

When in list mode (journal items) -  there could be hundreds of calls to the server.

This commit attempts to batch these calls