[FIX] crm: fix sales team embed kanban
Purpose:
--------

In sales team, the gauge showing the count of assigned leads in the members
kanban cards was not shown (only a number was displayed, the gauge had its
width and height set to 0).

Style has been added in the field's options so that the gauge has the same
dimensions as in 15.3.