[POC] don't add dependency to pytest nor upgrade pylint
The assert method could be improved as well as adding a setup to
define the linter and spliting the test, but this is a way
to make test independant from pytest and maybe a little more odoo
style using standard assert and minimal tools.

We may also have a simple utility that return the messages and assert
equal/false directly.