[FIX] mail: `followButtonView`, review design
This commit sets the button's width to the longest string, either
'Following' or 'Unfollow', depending by the translation.

It also force unfollow action color since it wasn't using the "warning"
one before.
Since it appears on :hover only, the text color provided to the user was
the :hover(ed) version that's substantially darker.

task-3005130