[IMP] web_editor: prevent table handle in contentEditable false
When a table isContentEditable=false:
- Don't show the resize cursor
- Don't show the table ui menu (which allows to move columns/rows)

Currently in Knowledge, when an embedded view contains a `<table>` element, was
is possible to use the editor to edit rows and columns of the view and it is not
intended. Those table are not editable, and as such the menu to edit them should
not be shown.

Task-3012552 and Task-3000318