[FIX] account: handle writing just tax_ids on an aml
account_avatax sets tax_ids in account.move's _compute_avalara_taxes()
function:

line.tax_ids = detail['tax_ids']

In certain cases tax lines (lines with tax_line_id) were not
generated. This resulted in errors about the journal entry being
unbalanced.

The tax line is supposed to be generated by account.move's
_sync_dynamic_line as follows:

Before the yield dirty lines (lines with a True compute_all_tax_dirty
field) are recorded and are then marked as not being dirty. During the
yield tax_ids is updated. This is a dependency of _compute_all_tax
which will cause compute_all_tax_dirty to be set to True again. When
the yield in the _sync_dynamic_line contextmanager returns we again
search for dirty lines, if dirty lines exist this early return is
bypassed and tax lines are generated:

if dirty_recs_before and not dirty_recs_after:
   return

However, when writing tax_ids on a single line we end up with
following nested _sync_dynamic_line contextmanagers:

1. old taxes and wrote false on compute_all_tax_dirty
     ...
2.   new taxes and wrote false on compute_all_tax_dirty
3. taxes changed and compute_all_tax_dirty is [False, False]

The contextmanager on line 1 is supposed to generate the tax
line. Line 1 shows that before the yield it sees the original
tax_ids (before the write), It sets the dirty field of the lines to
False. The problem occurs on line 2: a _sync_dynamic_line call has
occurred via the self.line_ids._sync_invoice() call in
_sync_dynamic_lines, it does:

line.balance = amount_currency

This write on account.move.line calls _sync_dynamic_lines again. When
_sync_dynamic_lines calls a new _sync_dynamic_line the dirty field is
True, because the write on tax_ids has now happened. It sets it to
False. When the yield of the line 1 contextmanager returns the dirty
field remains False and the it early returns, causing no tax line to
be generated.

No existing tests fail because the issue doesn't occur when writing on
multiple lines at the same time. This also explains why the issue can
not be reproduced manually in the webclient by setting a tax_ids on a
line; the webclient will always send the tax_ids of all the lines,
even when just changing one:

...
[
  [1, 967, {"tax_ids": [[6, false, [1]]]}]
  [4, 968, false]
]
...

The test_setting_tax_separately was added to cover this case.

The _disable_recursion contextmanager will prevent nested calls like
the one on line 2, but it did not work because it sets context on the
container, not self. To fix the issue stop using self and instead use
the records in the container.

opw-3016530