[IMP] bus: allow shared worker script to be updated
Before this commit, the shared worker script was cached by the browser
and was not systematically updated when the script changed. Indeed, the
shared worker is kept alive by remaining tabs and is not updated while
at least one tab is active.

This commit fixes this issue by adding the websocket worker version to
the script url: if the version changes, a new `SharedWorker` is used, with
the last script available. The others keep the old one until the next
reload of the tab.