[FIX] search & replace: next scroll not working on single search
Search for a single word which is outside of viewport, sheet will auto
scroll to that word, now again scroll back in such a way that the cell
is out of the viewport, now even if you press enter or click on
next / previous in searcg panel nothing will happen.
This behavior is changed in this commit and now it again scrolls you
back to that cell.

Task-3018173