[FIX] web_editor, website, *: keep old assets active
*: mass_mailing, test_assetsbundle, website_blog, website_mail_group,
   website_mass_mailing, website_payment, test_assetsbundle

When updating modules, old assets are made inactive. This is a problem
if they are still used.

This commit makes front end assets `noupdate` so that their `active`
state is not updated.
We rely on the `website_disable_unused_snippets_assets` weekly cron to
disable them afterwards if they are not used anymore.

task-2963840