[IMP] stock: print all labels for a picking
When printing labels of a picking, it "Transfer Quantities" is selected
(it is the default choice), it will print the product's labels based on
the done quantity of each product.
This raises an issue when trying to print labels from a "Ready" picking,
as the done quantities will be still at 0, which then prints blank
sheets.

While it is possible to set a custom quantity, this doesn't allow to
select different quantities for the different products. This commits
aims to solve this issue by allowing to print the labels regardless of
their done quantity but only based on the initial demand in the picking.

Part of task-2985735